# Fix performance issue introduced by vendor code