@Taowyoo 2023-06-28
Or maybe I need to export a C define from `mbedtls-sys`, such as `__SGX__`. And use this to conditional compile the change above