@arai-fortanix 2023-06-28
I haven't looked at this in full detail yet, but one thing I'm not thrilled about is directly conditioning the code on `__SGX__` being defined. The existing code seems to be pretty good at defining macros for features/properties like `MBEDTLS_EFFICIENT_UNALIGNED_ACCESS` and then conditionalizing the code on those macros. The feature macros get set based on things like the defines for the CPU architecture, so there's a level of indirection there.

This is a nice thing to do because at some point in the future there might be some other non-SGX platform that wants to use these alternatives for memory access. And it would be weird if they had to define __SGX__ to get the behavior they want, since since that platform wouldn't have anything to do with SGX.

So I think you want to conditionalize the code on a macro named something like `MEMCPY_IS_SLOW` or something like that, and we can set the `MEMCPY_IS_SLOW` macro if `__SGX__` is set. Then later someone else can compile with `MEMCPY_IS_SLOW` or make their own change to the header to automatically turn on `MEMCPY_IS_SLOW` for their platform as appropriate.

Minor C language thing: identifiers starting with double underscores are reserved for use by the platform. So things like the compiler or the system headers can define idenfiers that start with double underscores. Since this is something that we're defining, it shouldn't use double underscores. The standard library also uses some identifiers with single underscores, so those should be avoided as well.