@Taowyoo 2023-06-28
HI @arai-fortanix , 
Yes, it seems the upstream code added the `MBEDTLS_EFFICIENT_UNALIGNED_ACCESS` and apply such check on `mbedtls_xor` but not other macros in `alignments.h`, 

One possible solution here is to also add similar checks on `MBEDTLS_EFFICIENT_UNALIGNED_ACCESS` for family of `MBEDTLS_PUT_UINT32_BE` macros. And this changes  are reasonable and could be up-streamed.