@jethrogb 2023-06-28
This PR doesn't make sense to me. The code is clearly intended to get optimized so that there are no actual calls to memcpy. If our generated code still has calls to memcpy, we need to fix how the code is compiled.

Having compiled both Linux and SGX in release mode:

```
$ objdump -j .text.mbedtls_internal_sha256_process -r target/x86_64-fortanix-unknown-sgx/release/build/mbedtls-sys-auto-f19242cb09f8b77a/out/build/library/CMakeFiles/mbedcrypto.dir/sha256.c.obj

target/x86_64-fortanix-unknown-sgx/release/build/mbedtls-sys-auto-f19242cb09f8b77a/out/build/library/CMakeFiles/mbedcrypto.dir/sha256.c.obj:     file format elf64-x86-64

RELOCATION RECORDS FOR [.text.mbedtls_internal_sha256_process]:
OFFSET           TYPE              VALUE 
0000000000000080 R_X86_64_PLT32    memcpy-0x0000000000000004
000000000000009a R_X86_64_PC32     .rodata.K-0x0000000000000004
0000000000000b1e R_X86_64_PLT32    mbedtls_platform_zeroize-0x0000000000000004
```

```
$ objdump -j .text.mbedtls_internal_sha256_process -r target/release/build/mbedtls-sys-auto-030256f45744a275/out/build/library/CMakeFiles/mbedcrypto.dir/sha256.c.o

target/release/build/mbedtls-sys-auto-030256f45744a275/out/build/library/CMakeFiles/mbedcrypto.dir/sha256.c.o:     file format elf64-x86-64

RELOCATION RECORDS FOR [.text.mbedtls_internal_sha256_process]:
OFFSET           TYPE              VALUE 
0000000000000082 R_X86_64_PC32     .rodata.K-0x0000000000000004
0000000000000b06 R_X86_64_PLT32    mbedtls_platform_zeroize-0x0000000000000004
0000000000000b32 R_X86_64_PLT32    __stack_chk_fail-0x0000000000000004
```

So it's clear that the compiler fails to properly optimize out the call to memcpy. The fix should be in the compilation, not the code.