@Taowyoo 2023-06-29
So the reason why `memcpy` is not been optimized is:
When compiling, if build script of `mbedtls-sys` detect the environment is `freestanding` , see `mbedtls-sys/build/features.rs` , it will add cflag: `-ffreestanding` which equals `-fno-builtin` .

Then with  `-fno-builtin`  , `clang` will **Disable implicit builtin knowledge of functions** which cause `memcpy` and other builtin functions not been optimized out.

If assuming `-fno-builtin`  is necessary , then there are two solutions:
1. We should define exactly which builtin functions should be disable one by one through `-fno-builtin-<value>` to ensure only necessary functions are not been optimized.
2. According to https://man7.org/linux/man-pages/man1/gcc.1.html:

    > With the -fno-builtin-function option only the built-in
    > function function is disabled.  function must not begin with
    > __builtin_.  If a function is named that is not built-in in
    > this version of GCC, this option is ignored.  There is no
    > corresponding -fbuiltin-function option; if you wish to
    > enable built-in functions selectively when using -fno-builtin
    > or -ffreestanding, you may define macros such as:
    > 
    >         #define abs(n)          __builtin_abs ((n))
    >         #define strcpy(d, s)    __builtin_strcpy ((d), (s))

   We need to add `#define memcpy(dst, src, sz)  __builtin_memcpy((dst), (src), (sz))` in `mbedtls-sys/vendor/library/alignment.h` after `#include <string.h>` to ensure all calls to `memcpy` in this file will be optimized.