@jethrogb 2023-06-29
> The second solution is easy to do but need to change vendor code.

We can just add this as a separate include file that we pass to CMake, right? Or, similary, put it in the config.h.