@jethrogb 2023-06-29
Ideally we don't disable builtins that we provide in strcpy. But there's no way to instruct the compiler, ignore all builtins, except these specific ones?