@Taowyoo 2023-06-29
Hi @jethrogb , 
> But there's no way to instruct the compiler, "ignore all builtins, except these specific ones"?
As far as I know after some search on internet, the only way is to add something like ` #define abs(n)          __builtin_abs ((n))` in source code where you want to except some functions.

And as you can see this PR, I just updated, because of the limitation of C macro, this kind of C macros need to:
1. Add after include header file, so add this in `config.h` will cause compile error.
2. Such C macro cannot dealing with C code, if they call `memcpy`  like this: `memcpy(a, A_MACRO_EXPAND_TO_TWO_ARGUMENTS(b))`