@Taowyoo 2023-06-29
> I'm not thrilled by forcing an inclusion of string.h.

Me too. But this is only way how we could avoid changing vendor code.

I am thinking, is it possible to have a CMakeList.txt on top of `mbedtls` 's CMakeList.txt to overrride some vendor code without touching the `vendor` folder